+++ src/helpers/padding.js
@@ -1,5 +1,23 @@
 //@flow
-import composePositionAndSize from "./util/combinePositionAndSize";
+import { call } from "ramda";
+import Style from "further";
+
+import createHelpers from "./util/createHelpers";
+import combine from "./util/combine";
+import sizes from "./util/modifiers/sizes";
+import directions from "./util/modifiers/directions";
 import lens from "./util/lenses/baseSize";
 
-export default composePositionAndSize({ name: "padding", lens });
+export default call(
+  x =>
+    createHelpers(x, sizes).reduce(
+      combine,
+      createHelpers(x, directions)
+        .map(y => ({
+          ...y,
+          style: createHelpers(y, sizes).reduce(combine, y.style),
+        }))
+        .reduce(combine, x.style),
+    ),
+  { property: "padding", style: Style(lens("padding")) },
+);