+++ src/helpers/rounded.js
@@ -4,20 +4,18 @@ import Style from "further";
 
 import createHelpers from "./util/createHelpers";
 import combine from "./util/combine";
+import convertToBoxModel from "./util/convertToBoxModel";
 import sizes from "./util/modifiers/sizes";
 import directions from "./util/modifiers/directions";
-import lens from "./util/lenses/baseSize";
+import lens from "./util/lenses/baseBorder";
 
 export default call(
   x =>
     createHelpers(x, sizes).reduce(
       combine,
       createHelpers(x, directions)
-        .map(y => ({
-          ...y,
-          style: createHelpers(y, sizes).reduce(combine, y.style),
-        }))
+        .map(convertToBoxModel)
         .reduce(combine, x.style),
     ),
-  { property: "margin", style: Style(lens("margin")) },
+  { property: "borderRadius", style: Style(lens("borderRadius")) },
 );