+++ src/helpers/util/lowerCaseFirstLetter.js
@@ -1,4 +1,0 @@
-//@flow
-export default (str: string): string => {
-  return str.charAt(0).toLowerCase() + str.slice(1);
-};