+++ src/helpers/util/modifiers/__tests__/sizes.js
@@ -1,0 +1,10 @@
+import sizes from "../sizes";
+
+it("should match snapshot", () => {
+  expect(
+    sizes.map(({ reducer, ...x }) => ({
+      ...x,
+      result: reducer(8),
+    })),
+  ).toMatchSnapshot();
+});