+++ changelog.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [0.0.1] - UNRELEASED
 
 ### Added
-  - inital helpers from runkin (@jbaxleyiii)
+  - helper generation methods (@jbaxleyiii)
+  - position and rounded helpers (@jbaxleyiii via @delianides and @jakedawkins)
 
 ### Fixed